+++ src/pages/account-page/account-page.routes.ts
@@ -2,16 +2,13 @@ import { Routes } from '@angular/router';
 import { AccountPageComponent } from './account-page.component';
 
 export const AccountPageRoutes: Routes = [
+  { path: '', redirectTo: '/account/profile', pathMatch: 'full' },
   {
     path: '',
-    component: AccountPageComponent,
-    children: [
-      { path: '', redirectTo: '/account/profile', pathMatch: 'full' },
-      {
-        path: 'profile',
-        data: ['Profile'],
-        loadChildren: './profile-frame/profile-frame.module#ProfileFrameModule'
-      }
-    ]
+    component: AccountPageComponent
+  }, {
+    path: 'profile',
+    data: ['Profile'],
+    loadChildren: './profile-frame/profile-frame.module#ProfileFrameModule'
   }
 ];