+++ src/shared/helpers/http.helper.ts
@@ -21,6 +21,23 @@ export class HttpHelper {
     return this.authHttp.get(url, options);
   }
 
+  patch(url: string, data?: any): Observable<Response> {
+    if (data === undefined) {
+      data = {};
+    }
+    if (data.format !== undefined) {
+      data = data.format();
+    }
+    let bodyString = JSON.stringify(data);
+    let headers = new Headers({ 'Content-Type': 'application/json' });
+    let options = new RequestOptions({ headers: headers, withCredentials: this.withCredentials });
+    if (url.indexOf('mockapi.io') !== -1) {
+      options = new RequestOptions({ headers: headers });
+      return this.http.post(url, bodyString, options);
+    }
+    return this.authHttp.patch(url, bodyString, options);
+  }
+
   post(url: string, data?: any): Observable<Response> {
     if (data === undefined) {
       data = {};